test(quant): move QuantEmbedding export-path tests to tests/gpu/
_export_quantized_weight ends with torch.cuda.empty_cache() at
unified_export_hf.py:620, which raises on CPU-only CI runners that have a
CUDA-built torch wheel but no NVIDIA driver. The two export tests added in
the original embedding-quant change therefore failed on the linux unit job.

Move TestQuantEmbeddingExport (test_quantized_weight_is_packed_and_scales_registered
and test_tied_embedding_export_skips_packing) plus its _EmbeddingOnly /
_TiedEmbeddingLM / _embedding_nvfp4_cfg helpers from
tests/unit/torch/quantization/test_quant_embedding.py to a new
tests/gpu/torch/export/test_export_embedding.py. This matches the existing
convention -- tests/gpu/torch/export/test_export_weight_gpu.py already
imports _export_quantized_weight directly.

The remaining tests in tests/unit/.../test_quant_embedding.py only exercise
the forward / configuration paths and pass under CUDA_VISIBLE_DEVICES="".
